Created different scene classes
