+++ GameFramework/include/Scene.h
@@ -19,16 +19,9 @@ public:
     virtual void update(sf::Time deltaTime){}
     virtual void render(sf::RenderWindow& window){}
 //This should be protected
-//protected:
+protected:
     std::vector<std::unique_ptr<Entity>> m_Entities;
 };
 
-class MainMenu : public Scene{
-public:
-    MainMenu();
-    ~MainMenu(); 
-    void update(sf::Time deltaTime) override;
-    void render(sf::RenderWindow& window) override;
 
-};
