+++ Project/CMakeLists.txt
@@ -1,4 +1,3 @@
-
 add_executable(Project main.cpp)
 
 target_link_libraries(Project PUBLIC GameFramework)
