+++ Project/include/Play.h
@@ -1,0 +1,10 @@
+#pragma once
+#include "Scene.h"
+
+class Play : public Scene{
+public:
+    Play() = default;
+
+    void update(sf::Time deltaTime) override;
+    void render(sf::RenderWindow& window) override;
+};
